implement takeBestAction

+++ src/TicTacToe.hpp
@@ -25,7 +25,7 @@ class TicTacToe {
 	//	Take best action using minimax algorithm
 		void takeBestAction();
 		std::vector<byte>& possibleActions();
-		byte minimaxValue();
+		short minimaxValue();
 
 	//	Returns 0 if the game is still going, 1 if current player won, -1 if the opposite player won, and 2 if tie
 		char isGameOver();
